fin authorization

+++ app/Http/Controllers/HobbyTagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Tag;
 use App\Hobby;
+use Illuminate\Support\Facades\Gate;
 
 class HobbyTagController extends Controller
 {
@@ -21,6 +22,11 @@ class HobbyTagController extends Controller
     public function attachTag($hobby_id, $tag_id)
     {
         $hobby = Hobby::find($hobby_id);
+
+        if(Gate::denies('connect_hobbyTag', $hobby)){
+            abort(403, 'forbidden');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->attach($tag_id);
 
@@ -30,6 +36,11 @@ class HobbyTagController extends Controller
     public function detachTag($hobby_id, $tag_id)
     {
         $hobby = Hobby::find($hobby_id);
+
+        if(Gate::denies('connect_hobbyTag', $hobby)){
+            abort(403, 'forbidden');
+        }
+        
         $tag = Tag::find($tag_id);
         $hobby->tags()->detach($tag_id);
